reorganize code and components
