+++ src/App.jsx
@@ -1,14 +1,9 @@
 import React from "react";
 import "./App.css";
-import Router from "./Router/Router";
-
-
-
-
+import Router from "./Router";
 
 function App() {
   return <Router />;
 }
 
 export default App;
-
