+++ src/Components/auth/AuthForm/AuthForm.jsx
@@ -1,6 +1,5 @@
 import React, { useRef } from "react";
-import useApi from "../../hooks/use-api";
-
+import useApi from "../../../hooks/use-api";
 
 export default function AuthForm({ handleSubmit }) {
   const { isLoading, error, online } = useApi();
@@ -18,7 +17,6 @@ export default function AuthForm({ handleSubmit }) {
   return (
     <div>
       <div>
-        
         <form action="get" onSubmit={onSubmit}>
           <input type="text" name="email" ref={email} />
           <input type="text" name="password" ref={password} />
