+++ src/Pages/Board1.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import Board1 from "../components/boards/Board1";
+
+const Board1s = () => {
+  return <Board1></Board1>;
+};
+
+export default Board1s;
