--- src/Pages/Board1/Board1.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-import Board1 from '../../Boards/Board1'
-
-const Board1s = () => {
-  return (
-    <Board1></Board1>
-  )
-}
-
-
-export default Board1s
