+++ src/Pages/authentication/Login.jsx
@@ -1,22 +1,21 @@
-import React from 'react'
+import React from "react";
 
-import useApi from '../../hooks/use-api';
-import AuthForm from '../../Components/AuthForm/AuthForm';
-
+import useApi from "../../hooks/use-api";
+import AuthForm from "../../components/auth/AuthForm/AuthForm";
 
 export default function Login() {
   const { login } = useApi();
-  
+
   const handleLogin = (username, password) => {
     login(username, password).then((res) => {
       console.log(res);
-    })
+    });
   };
-  
+
   return (
     <div>
       <h1>Login</h1>
       <AuthForm handleSubmit={handleLogin} />
     </div>
-  )
+  );
 }
