+++ src/Pages/authentication/Register.jsx
@@ -1,15 +1,14 @@
 import React, { useRef } from "react";
-import AuthForm from "../../Components/AuthForm/AuthForm";
+import AuthForm from "../../components/auth/AuthForm/AuthForm";
 import useApi from "../../hooks/use-api";
 
-
 export const Register = () => {
-  const {register} = useApi();
+  const { register } = useApi();
 
   const handleRegister = async (e) => {
     register(username, password).then((res) => {
       console.log(res);
-    })
+    });
   };
 
   return (
@@ -17,5 +16,5 @@ export const Register = () => {
       <h1>Register</h1>
       <AuthForm handleSubmit={handleRegister} />
     </div>
-  )
+  );
 };
